feat(component): refactor PopularCarousel into smaller reusable components
